requirements.txt
Flask==2.2.2
flask_login==0.6.2
flask_sqlalchemy==2.5.1
python-dotenv==0.21.0
PyYAML==6.0
Werkzeug==2.2.2
